backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
pytest==8.3.3
pytest-cov==6.0.0
google-genai
google-api-core>=2.19.0
httpx==0.27.2
